requirements.txt
flask
flaks-wtf
flask-sqlalchemy
flask-migrate
flask-login
email-validator
schedule